src/requirements.txt
alembic==1.10.2
bcrypt==4.0.1
bleach==6.0.0
blinker==1.5
boto3==1.26.95
botocore==1.29.95
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
mkdocs-material-extensions==1.1.1
passlib==1.7.4
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.7
SQLAlchemy-Utils==0.40.0
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1